changed matching directory to like directory in workers directory
